MDL-77216 editor_tiny: Moved get_string from JS to PHP
We need to call get_string with async inside the init process to remove a sub-menu item,
but since the editor init is a critical code path and must not be blocked by any synchronous task,
We have to change to get the string. This patch is moving the get_string function from JS to PHP,
and the patch also provides the input in an array so we can remove multi submenu items in one call.