MDL-78850 calendar: Added course parameter in the calendar view url
There are four changes that need to be made:

1. The view calendar URL needs to include the course ID param so that when a user clicks on a date or
the full calendar inside the course page's calendar block, the course calendar will be displayed.

2. The course name in the course calendar should be clickable to make it easier for users to navigate to the course page.

3. When users create a new event, the course field should automatically select the course where
the user is regardless of the course they have selected in the course dropdown at the calendar view.
But, this is just a default and users can change it if they want.

Co-authored-by: santoshndev <santosh.nag2217@gmail.com>