MDL-78862 theme: Ensure the file is readable before renaming it
If the copy() function returns true, it indicates that the file
has been successfully copied.
However, based on the log reported on MDL-78862,
this may not be completely accurate.
It's possible that the entire copy process has not yet finished,
but PHP has already returned the true value.
In such a case, a limited loop will be implemented to
verify if the copied file exists and is permitted to
read so it can be renamed.